requirements_test.txt
pytest-homeassistant-custom-component
mock
xmltodict
pytest-asyncio
packaging
